+++ src/Graphdoc/Analysis.hs
@@ -1,0 +1,11 @@
+module Graphdoc.Analysis
+  ( module Graphdoc.Analysis.Opt
+  , module Graphdoc.Analysis.CLI
+  , runAnalysis ) where
+
+import Graphdoc.Analysis.Opt
+import Graphdoc.Analysis.CLI
+
+runAnalysis :: Opt -> IO ()
+runAnalysis opt = do
+  putStrLn "hello"