+++ src/Treedoc.hs
@@ -1,8 +1,0 @@
-module Treedoc
-  ( module Treedoc.App
-  , module Treedoc.Readers
-  , module Treedoc.Writers ) where
-
-import Treedoc.App
-import Treedoc.Readers
-import Treedoc.Writers