+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for treedoc
 
+## 0.1.0.1 -- 2020-12-16
+
+* Can poorly convert from generic markup trees to Texinfo
+  trees. Soon moving on to new architecture with two parts.
+
 ## 0.1.0.0 -- 2020-10-16
 
 * First version. Can read a directory of markup files, convert them, and write